SDC-9267: Negative number of output records displayed when SQL inserts fail on a bad datafile
It took me a while to get a proper repro case, luckily Mark's data on
the JIRA turned out to be very helpful. The issue surfaced when not all
the records had all the columns that are available in the table. This
lead into a situation when multiple statements were generated and then
when they failed, we have simply dumped the whole batch to error stream
for each failing statement - so if two statements failed, we have dumped
the whole batch twice.

This patch changes the logic a bit - now we're tracking what exact
records are associated with each statement and if  given statements
fails, we only propagate to error the associated records.

Change-Id: I940f324c71b6f3be84f259ff799194b1242a03a7
Reviewed-on: https://review.streamsets.net/15112
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Ji Sun Kim <jisun@streamsets.com>