SDC-9320: Cluster YARN Streaming from Kerberized Kafka throws IllegalArgumentException
This happened because the origin is "self committing" and we're trying
to commit the idle batch that does not have any real offset. Thus this
patch adds an explicit flag to the FullPipeBatch about the origin of the
batch so that the committing code know that it's safe to ignore.

Change-Id: Idd329f2ed06c47db1fe34eff4daccf1d1a4d250d
Reviewed-on: https://review.streamsets.net/15178
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Hari Shreedharan <hshreedharan@streamsets.com>