SDC-9359. JDBC producer does not check if record is empty
Make JDBC producers log a warning and then ignore records which have zero parameters

Note: multi-row record writer was previously throwing a different (specific) exception with error
JDBC_22 in the case where the set of parameters was empty (i.e. from an empty record).  In order
to follow the same behavior as the generic (i.e. single row) record writer, that behavior was also
changed here to log a warning and ignore the record.  This change in behavior is done to make both
kinds (single or multi-row) have the same handling for empty records, which would likely be consistent
with what a user would expect

Adding tests

Change-Id: I5ec4431a733d9709942593fb0242fa00bc084c0e
Reviewed-on: https://review.streamsets.net/15255
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>