SDC-9408: Re-add pluginRepoistory for CDH
I've recently removed duplicate repository configuration across all the
different CDH stage libraries (SDC-9384). And I've also removed plugin
repository for the CSD as it pointed to exactly same repo. All seemed
working fine when building on maven 3.3, but from some reason this
breaks on maven 3.5. My best guess is that the older maven version used
the normal dependency repositories to download plugins as well, but the
newer version does not and hence the build failure.

This patch simply adds the plugin repository back. But rather then
putting to random child pom files, I've put it to the root pom so all
repositories are in a single place.

Change-Id: I95504d0d9d0eacd8e5aacd4c8e108e47bd9e50c6
Reviewed-on: https://review.streamsets.net/15317
Reviewed-by: Ji Sun Kim <jisun@streamsets.com>