SDC-9425: Attempt to capture process PID in ShellExecutor
This patch updates the shell executor to attempt getting process id with
reflection. I do understand that the reflection is brittle and can break
on the next JDK release - that is why the main behavior does not depend
on this functionality at all. It's used only for debugging purposes when
for example the customer need to see what process we have started versus
what process is running.

Change-Id: I7c3d447220de6f68c20fc97c4e402dfbdde5464c
Reviewed-on: https://review.streamsets.net/15346
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>