SDC-9453: Load configuration pushed down by control hub
This is a first step on having ability for control hub to push down
configuration to data collector. This patch provides method to store
control hub pushed properties  and a way to load it into the main
configuration upon start.

The facility is not wired to the rest of the data collector - this will
happen in subsequent JIRA.

Change-Id: If5ac73cd26f82a2c4c94829bb87a5167826848b4
Reviewed-on: https://review.streamsets.net/15408
Reviewed-by: Jeff Evans <jeff@streamsets.com>