SDC-9459: StageLibraryTask needs to keep track of available stage libraries
I will need ability to see list  of  loaded stage libraries for a future
use case, but  the patch seems "standalone", so I wanted to submit it
separately. It also solves the trouble on Package Manager where the
manager was only "extrapolating" what  stage libraries are available
rather then getting that list explicitly.

Change-Id: Ic05cd05c552da2c1fa05292919335972aaedb25f
Reviewed-on: https://review.streamsets.net/15415
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>