SDC-9479: Push shared managers to AbstractRunner rather then re-defining them in the child classes
Most of the managers (pipelines store, state store, ...) are defined in
all the runners and in the parent class as well - and because of that we
have the same method with the same implementation in three different
places.

This patch refactors that to have the managers in a single place.

Change-Id: Ic13cdd1b6c7decfb942173de1a698f5d3abee681
Reviewed-on: https://review.streamsets.net/15491
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Jeff Evans <jeff@streamsets.com>