SDC-9484: Move generation of attributes to a single place in the Runner infrastructur
The way we generate attributes that are then stored in the state file
is distributed across several files and it's hard to follow and very
error-prone even today. For example, runtime parameters are mistakenly
stored twice or they are not properly stored for STARTING state.

This patch moves responsibility for state attributes only to the Runnner
implementation with shared logic available in AbstractRunner to have as
few places with all the details as feasible.

Change-Id: Ia1a22175525d55fb5d8b6d0576c77c549999e687
Reviewed-on: https://review.streamsets.net/15501
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Jeff Evans <jeff@streamsets.com>