SDC-9519: Validation exception -- Amazon S3 destination for Avro data format
When using Avro format, the kafka schema registry client dependency is
must have - regardless whether one actually uses the client or not. The
dependency is required to even load our classes. From some reason the
main commonlib protolib declares this dependency only "optional" so that
each and every stage library needs to explicitly depend on it. I assume
that we chose this to have ability to not ship this dependency in stage
libraries that do not need Avro at all.

For now I've added the explicit dependency to dataformat stage library -
once we fully transition everything to services model, the commonlib as
it exists today will cease to exist anyway, so this problem will
naturally go away.

Change-Id: I86a56fcc66ac17c70bb7a66beeaa7fbe885989eb
Reviewed-on: https://review.streamsets.net/15610
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Anthony McKeehan <anthony@streamsets.com>