SDC-9575: SCH Registration: Update CSD to use the new command line tool
Final piece of the puzzle changing the actual Cloudera Manager
integration called CSD.

This patch is a backward incompatible in a sense that newer CSD will
not be able to register token for older parcel. However that is sort of
intentional - the existing scripts were so flaky that I would prefer
them gone and rather fail fast.

Change-Id: I30fcfc85351e7eb8bb6551d6a139d107f224bd53
Reviewed-on: https://review.streamsets.net/15719
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Jeff Evans <jeff@streamsets.com>