SDC-9593: NPE when creating multithreaded pipelines with services
The trouble is that for multi-threaded pipelines we need to duplicate
pipeline runner (all stages inside the pipeline canvas minus the
origin). We don't do that on pipeline creation as at that time we don't
know how many runners should exists and instead we do that on pipeline
init() - hence the code for duplication is slightly different then for
creation.

This code was incorrectly creating definition map from the service
implementation class name rather then from the interface that the class
exports and thus the right service was never found leading to this NPE.

Change-Id: I193f0f55781543591ebb683ccb1680f9a32ef770
Reviewed-on: https://review.streamsets.net/15759
Reviewed-by: Keith Burns <keith@streamsets.com>