SDC-9609: some KAFKA_37 errors should stop the pipeline.
The underlying trouble is that our dataformat library is catching
Throwable - e.g. both "Exception" as well as "Error" classes. I don't
think that it's right for the dataformat library to catch Errors as in
this case, it was OutOfMemoryError that was caught and "swallowed" -
which definitely looks wrong.

Change-Id: Iaf3300440a27b5b2d5fe8b4e4de596b727be8754
Reviewed-on: https://review.streamsets.net/15784
Reviewed-by: Bob Plotts <bob@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
(cherry picked from commit 2d715adde4546d1bcf1407023707b4ffaa0f2627)
Reviewed-on: https://review.streamsets.net/16059
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>