SDC-9601: Provide REST interface for validating detach stages
This patch adds support for static validation of detached stages in the
REST interface. It works very similarly as the Pipeline store rest
interface.

Doing actual 'preview' (e.g. running the stage init()) is not part of
this patch and will be done at same point in the future.

The detached stages are not persisted, the REST interface will take the
detached stage, upgrades it, validate it and return back the updated
version.

Change-Id: I9e6f1afdb3f9800449395c4a8e7471043c6e6751
Reviewed-on: https://review.streamsets.net/15793
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Anthony McKeehan <anthony@streamsets.com>