SDC-9413. Possibility of data loss (files could be missed) in case when files are being uploaded in pipeline's source bucket and simultaneously pipeline is also running and copying files from source to destination bucket.
There is a race condition when the listObjectsChronologically() is scanning the list of files in S3.
Let say we have {1.txt, 2.txt, 3.txt} in S3 and S3 API returns the list of files in batch size = 2.
* The first batch returned, {1.txt, 2.txt}, meanwhile, 4.txt is added and then and 0.txt is added.
* The second batch returns, {3.txt, 4.txt} since 0.txt is lexicographically less then 2.txt, 0.txt is ignored on the second batch call.
* So, we have {1.txt, 2.txt, 3.txt, 4.txt} in the queue, we produce the records and save the last offset of 4.txt
* the queue is empty, we start to find the list of files in S3
* 0.txt will be ignored because it is less then 4.txt (last source offset)

Once we call the S3 API to get the list of files and sort them in last modified timestamp. We should not call by the batch size because of the above edge case.

Change-Id: I222f4fadb0f97cf60d407eccf60f22302c9f2928
Reviewed-on: https://review.streamsets.net/15882
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>