SDC-9673: SDC UI displays logs for unrelated pipelines
I believe that I was able to reproduce this issue and track it down.
When sending remote event for starting a pipeline, we set a lot of
variables in MDC for logging purpose, but we actually never clean it up.
So the same thread, if reused, will still contain the MDC configuration
from last pipeline start command.

This patch simply cleans up MDC when done with starting remote pipeline
so that the event handler thread is "pristine".

Change-Id: I580e90d4d64705a73f8ea987c6db7329231b1584
Reviewed-on: https://review.streamsets.net/16194
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>