SDC-9845: Remove PURE_FUNCTION flag from Static Lookup
I've initially added the PURE_FUNCTION flag on the static
lookup processor as it is indeed a pure function that doesn't
depend on the environment. I've however seen it used in
environments where the user is just using static look up to
see how other lookup processors would behave and thus it will
be more beneficial to not optimize this specific processor.

Change-Id: I66155e6b1538aa803ea3b4df4f2b27f360756627
Reviewed-on: https://review.streamsets.net/16319
Reviewed-by: Anthony McKeehan <anthony@streamsets.com>