SDC-9904: Registered SDC must be restarted for DP to function
This is my oversight - my original version of SDC-9453 made the method
RuntimeInfo.storeControlHubConfigs do two jobs - 1) store the new
configs from control hub in a persistent store and 2) update runtime
configuration. Through review this logic was changed and the given
method is now only responsible for 1).

As it happens we can set the control hub propagated configuration in two
places (in two different ways) and only one of the ways was properly
updated to do 2).

Change-Id: I89a119b18d2feb1432f83a5fd940ae73deb487b6
Reviewed-on: https://review.streamsets.net/16478
Reviewed-by: Anthony McKeehan <anthony@streamsets.com>