SDC-9925. ADLS destination: NPE when Max File Size or Max Records is set.
This patch fixes the ADLs Destination throwing NPE when max file size/max records are set. The error happens when we don't have a file open by the writer and are checking for whether we need to roll file, but without an counting output stream initialized, we are getting a NPE when we are trying to get byte count for the stream. This patch guards against that.

Change-Id: Ifcd83099d9a9217103759421e00857da2db6d1b2
Reviewed-on: https://review.streamsets.net/16520
Reviewed-by: Ji Sun Kim <jisun@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
(cherry picked from commit dfd700fadcb75c2bf8b276d7eda7c31b454d80bc)
Reviewed-on: https://review.streamsets.net/16550
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>