SDC-9938. Azure DataLake fails while producing close events when max record setting is enabled.
It looks like we have been calling both writer.close and clearStatus of which both apis were trying to rename the file, even though one succeeds another will fail (but the return status was a boolean), so we were trying to generate two close events. This is also because we were not clearing filePathCount state once we commit the file. This patch changes the logic to remove the dirPath from filePathCount when we commit and the dirPath will get automatically added the next time a file is created.

Change-Id: I42bb0dda6fa346ade709e254c2ad67c06285b8b7
Reviewed-on: https://review.streamsets.net/16522
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>
(cherry picked from commit 0cb017912b7586ce5734f8da790abde4108c592c)
Reviewed-on: https://review.streamsets.net/16551
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>