SDC-9945: Events are not protected
This turned into slightly bigger patch that I've anticipated - the core
of the change is moving the way interceptors are passed down from being
owned by the pipes to being owned by the FullPipeBatch. Honestly this
part of the code was the first few patches that I have written for the
interceptor integration and thus it's more of a clean up after seeing
how it fits with the rest of the infrastructure.

Nevertheless this patch changes both event and error sinks to properly
call all interceptors on their data - something that we were until now
not doing properly.

Change-Id: Ia5f91921727eaa204a82a53fcc29be989e9858fd
Reviewed-on: https://review.streamsets.net/16528
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>