SDC-9993. Files created on ADLS destination does not fully honor Max File Size setting.
This patch mainly aims at fixing the bug wherein the max file size is not completely honored by the ADLS destination. One of the main reasons for is that we were closing all files even though one file has reached the max file limit (this is the main reason for having files which are less in size), as a part of this patch also did a bit of refactoring to make sure we are calling classes/functions in the context of current thread and have the indirections between classes in the correct structure.

Change-Id: Iacee552fa0ea072452ffafe9554143a979e297fe
Reviewed-on: https://review.streamsets.net/16627
Reviewed-by: Ji Sun Kim <jisun@streamsets.com>