SDC-10001. Rollback to savepoints should not create records for rolled back statements.
Logminer generates one rollback statement per change. So if there were 2 updates issued
to a given row id after the savepoint was created, a rollback to the savepoint creates
two statements which revert the change made to the row in the reverse order.

So for each statement generated due to a rollback, we remove the last seen record
for that row id. This also works fine if the transaction was partially sent down the pipe,
since we don't remove or complain if a change does not exist (for example 2 rollback-generated
statements to one change statement is fine - we just don't do anything for the 2nd rollback
statement).

Unfortunately, this is a bit less efficient than I'd like, but this is the best we can do.
We can get the row id only once we parse the actual statement, and for each rollback-generated
statement we end up having to traverse the entire transaction in reverse till we find the associated
change.

Change-Id: I0b1ad8fcc94f74291bb6d25007c30f6df8daa65b
Reviewed-on: https://review.streamsets.net/16633
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Junko Urata <junko@streamsets.com>