SDC-10084: When numerous stages establish a connection(i.e. Salesfore) and have an error they should specify which stage is causing the issue
This is a rip-off or a similar change that I did when working on
multi-threaded pipelines - SDC-4927. Just instead of adding runner-id to
the logs, it adds the stage name.

I've decided not to propagate the new column to the UI as that is
already extremely wide and not really that helpful.

Change-Id: I2c2c54e4f5e1fb34ef87abfeae1def66231b3c86
Reviewed-on: https://review.streamsets.net/16875
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Jeff Evans <jeff@streamsets.com>