SDC-10148. Improve HBase Destination Performance.
Move table open and close code out of the procesing loop to
the init() routine. This way we open the table once.

Added a change to in init() to stop if we have an
error connecting to the table.

Added destroy() routine to close the table when finished.

Change-Id: Iceb959b94070a29c3933c0f8c0443a204b0d1c2f
Reviewed-on: https://review.streamsets.net/16929
Reviewed-by: Virag Kothari <virag@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
(cherry picked from commit 178d959bcedc650908fe4cbc8ba80e097d6f290d)
Reviewed-on: https://review.streamsets.net/17650
Reviewed-by: Bob Plotts <bob@streamsets.com>