SDC-10162. MapR Streams Multi-Topic consumer - batch size is ignored.
This patch is for testing and evaluation.

MapR sometimes returns 150 records per call, sometime other numbers
of records regardless of the limit which is set when calling the API.
In turn, the code in the Multi-topic Consumer simply returns those records
as a batch regardless of the batch size set in the Origin's UI.

Also, fixed assumptions in tests.

Change-Id: Ie7481b21970945ff87106ccb840298c090c561cd
Reviewed-on: https://review.streamsets.net/16981
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Anthony McKeehan <anthony@streamsets.com>