SDC-10242: Depend on compatible versions of mockito and powermock
Mockito and powermock versions work with tandem - with limited
compatibility set. Most of CDH stage tests are failing because the JDBC
driver brings yet another version of mockito through the artifact
mockito-all that sadly does not conflict with mockito-core that we
explicitly depend on.

There are still additional failures after this fix, but they no longer
see as mockito related.

Change-Id: I11f1dd2ec6d5799c257d4d1dcbe6a1953ff79be0
Reviewed-on: https://review.streamsets.net/17165
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>