SDC-10243: Spark tests failing with class signature
I have to say that I don't fully understand this one. CDH 5.12 and later
fails on bunch of random places with various class loading issues
(different class found then expected). When looking into CDh 5.11 the
only notable difference is addition of azure in CDH 5.12.

Interestingly enough there is a note at the begging of the fragment that
the azure libraries must appear after hadoop which wasn't the case. I
followed the advice and moved them down - and voi-la things started
working.

I assume that the Azure libraries are having a lot of dependencies
inside their jar file that mess with our tests.

Change-Id: I45b02ae05c1afa8be35a504c913fd9e9983ef581
Reviewed-on: https://review.streamsets.net/17176
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Ji Sun Kim <jisun@streamsets.com>