SDC-10344: SDC Security Manager should not block access to the blobstore
This turned out to be quite a big issue actually - the BlobStore has
been directly passed to stages that runs with different classloader -
thus there were moments when the BlobStore wasn't properly running with
it's own class loader which could lead to many random exceptions.

This patch introduces a Runtime wrapper like we have for everything else
that is crossing classloader boundaries.

Change-Id: I8606cbd20d23297194394b1b7f524e57d87a6b24
Reviewed-on: https://review.streamsets.net/17427
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Jeff Evans <jeff@streamsets.com>