SDC-7584: get logs REST api returns partial data
The trouble is that our logic is randomly seeking to the midle of the
logfile and is not doing offset correction to get at the end of the line
to start reading log messages. This patch updates that logic to ignore
first line(s) until a valid log line is found.

Change-Id: I105653b5fc025c80957a4d8c15c612a9eb9db8cc
Reviewed-on: https://review.streamsets.net/17521
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>