SDC-10376: Interceptors for stage 'AmazonS3_02' already registered
It turns out that preview is calling skipStage() method even non-stage
pipes which itself is weird, but it seems required for the logic (e.g.
the method is probably ill-named).

Calling the skipStage() for non-stage pipes meant that it will end up
being called more then once which we do not allow. Thus this patch sets
the interceptors if and only if the given pipe is a stage pipe.

Change-Id: I173744d80c244e6cda5f4514676499965e6e3a5c
Reviewed-on: https://review.streamsets.net/17523
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Anthony McKeehan <anthony@streamsets.com>