SDC-10512: Remove PURE_FUNCTION from expression evaluator behavior flags
The expression evaluator is an interesting processor that is very
powerful - sometimes used to just massage fields, but it can be used in
powerful way to even go to remote systems or load data from disk through
the various ELs. Thus it's not really always a PURE_FUNCTION albeit it
sometimes can be used in that way.

Change-Id: I8f8a3f81cfa9d43ce06581626cdedbb88977bca1
Reviewed-on: https://review.streamsets.net/c/18085
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Jeff Evans <jeff@streamsets.com>