SDC-10609: sdc_2.2.0.0_pipeline_Orgin_Processor.json
The aliases configured in default sdc.properties are no longer fully
valid. The destination stage library is actually not that easy as there
is no 1:1 mapping between CDH version and Spark 2.x version, so we
settle on the lowest available one.

Failing with wrong Spark stage library seems better then failing the
pipeline with random "class not found" exception.

Change-Id: I9aaae1412ca429e078484c46effbe087666451e4
Reviewed-on: https://review.streamsets.net/c/18617
Reviewed-by: Junko Urata <junko@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>