SDC-10708: SDC hangs with StackOverflowError when writing to CDH Kafka error destination with SSL
I can't explain why the Kafka test was failing as I'm convinced that JVM
should have already fully initialize the RuntimeInfo class which should
not lead to infinite loop.

But JVM works in mysterious ways and rather then chasing down why that
happen, I feel that it's better to simply make sure that the
SdcSecurityManager is not vulnerable to infinite recursion even when JVM
does something "unexpected".

Change-Id: I2030ff0692ddebf16c283a3fcccce2e7d14a1b9c
Reviewed-on: https://review.streamsets.net/c/18664
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>