SDC-10384: UI - Test origin - not reversible exception when previewing origin with event generation enabled
This is fundamentally just replacing one error for another one - instead
of failing with fairly cryptic validation error, we now throw a proper
exception describing what exactly is going on.

The problem is that when a pipeline have an origin with even handling
and we substitute it with test origin that doesn't support events at
all, the pipeline itself become invalid.

I was considering to simply prune the pipeline and remove stages from
the event path, but that seems a bit defeating the purpose of the test
stage - it doesn't actually test the pipeline properly. And thus I
decided to throw the clear exception instead.

There is still a case to be made that it would be generally useful if
the raw data origin could generate events - which is a story for another
JIRA.

Change-Id: I26ac2dddecd5f3f85e6d4f630445631c256cce97
Reviewed-on: https://review.streamsets.net/c/18868
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Jeff Evans <jeff@streamsets.com>