SDC-4188. Provide FTP/SFTP destination
This adds a new FTP/SFTP Destination which supports the Whole File format.  The majority of the new code is in the com.streamsets.pipeline.stage.destination.remote package, where the new Destination is defined.  However, a lot of refactoring was done in the com.streamsets.pipeline.stage.origin.remote package in order to share as much logic, configs, and verification code as possible between the source and new Destination (see the new com.streamsets.pipeline.lib.remote package).

Change-Id: Ie1a9a6fc3fa7f4ac005f97373af0cffcef89a0ee