SDC-10846. Improve handling of connection issues when reporting stats
It will retry 5 times, once a minute, before backing off for 1 day.  After 1 day, it will retry the 5 times again
and if it continues to fail, it will back off for 2 days.  After that, 4 days.  Finally, it will give up and
switch off.  This gives it about a week to be resolved before giving up, and prevents spamming the logs every
minute.

Change-Id: I14f6fd51a7563c9c475dee6e75c0f6cd1c606181
Reviewed-on: https://review.streamsets.net/c/19318
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>