SDC-10942: CSV parser misses fields when allowing extra columns
The core of the issue is that the 'extra column' counter has been
calculated on a per record basis. So in the test example the first
record had one extra column which means that we have extended the
headers by n=1. In the second iteration there were two extra columns,
but since the first extra column has already been registered, we run the
method with n=1 as well and created a duplicate header. Since the
example is creating LIST_MAP, duplicate headers are overridden and thus
not all columns are passed down.

This patch solves it by keep the extra column counter on the class
instance basis which is consistent with the headers list.

Change-Id: Ie8a21265f150aa698c83faaae3afd72331ef8edf