SDC-10925. basic-lib tests are hitting the timeout and are too slow due to not enough randomness
On machines without a lot of randomness, tests that use SecureRandom can block for minutes at a time; which
can add up.  We'll use a workaround to change random from /dev/random to /dev/urandom which does not block.

The timeout per-module for unit tests is 10 minutes.  That's not long enough for basic-lib on slower
machines where we've seen it take about 25 minutes.  We'll bump the timeout up to 30 minutes.

Change-Id: I0fb32411034fb7b53a37e20799be5552dbd41462