SDC-10987: JDBC Multitable Consumer multiple offset columns with initial offset
The problem isn't really with multiple offsets, but with setting the
offset in a way that the first batch reads nothing - can be trivially
reproduced with single key offset as well.

The problem is that were clearing the initial offset configuration
immediately after the first table context was created - regardless
whether that initial offset was actually used. This patch moves the
clearing logic to batch handling and resets it only if at least 1 record
was read from given table.

Change-Id: Ic96fde34430a8dad10242e52fbeb86a6aa32c00c