SDC-11036. X-Forwarded-For not honored when behind a proxy
This is part 1 of a two part issue. While there is a jetty ForwardedRequestCustomizer
that is added based on http.enable.forwarded.requests=true in
sdc.properties.

The SSOPrincipalUtils/HttpUtils classes do not honor this setting, and
always look for a CLIENT-IP header (which is non-standard, we dont
appear to use, or document). This change modifies that to use the
X-Forwarded-For header like jetty, but does not attempt to use the
sdc.properties setting to modify its behavior. That part I'm unclear how
to reconcile.

A related change is on the SCH side, as SCH does not expose this setting
in any properties file at all despite using the same WebServerTask
container.

I discovered this issue when the incorrect IP address was logged in the
Login Audit in SCH when behind one or more proxies.

Change-Id: I43da2d013619cf28f92ff16e8b4f576215ac90fa