SDC-11059: PreparedStatementCache should not close statements on cache eviction
When the class PreparedStatementCache evicts entries from the cache, they are
immediately closed. This is a problem because the entries (PreparedStatements)
are actually also cached in the main destination, so when the cache size is
less then what is needed, a user will see exception about working with closed
data set.

Change-Id: Ifeacc3f87801647c490883993862d5a552ee7e44