SDC-7934: test_azure_event_hub_producer test failure
This is an actual bug in the Azure Event Hub destination - we did not
explicitly called close() which means that file formats that are
serializing some pieces during closing the stream do not have chance to
do so - e.g. in this case JSON serialization as an array is serializing
the last ']' during the close() call.

Change-Id: I3bf29457fb3fcaaf85d06bb4d099a6d351ed1331