SDC-11160: Stop reordering records in JdbcGenericRecordWriter
This is evolution of the many issues we have seen when inserting CDC
data with the generic record writer re-ordering. Rather then keep
'improving' the re-ordering, this patch removes it completely and
re-uses only subsequent statements.

This way the destination will never re-order any queries. The approach
closely follow what we are already doing in JdbcMultiRowRecordWriter.

The patch is fairly sizable as many files had to be updated to take into
account removed cache (that is no more).

Change-Id: I79c2487508ab0ee3c4e3f7f2070a7de124fdf4c5