SDC-11170: Backend for package manager should not allow installing already installed stage library
The package manager for enterprise libraries isn't entirely ideal, but
there are few things that we can improve in the short term.

This path improves the logic in backend that will prevent one to install
stage library that is already installed which can happen in the UI if
one selects to install all available Snowflake versions for example.

Change-Id: I4b8f51ab9594f6794773f531be8f43a2af68288a