SDC-11238: Skip maven-dependency-check
The plugin seems to constantly fail our Jenkins build for somehow random
issues. Also it's not particularly useful as it's reporting issues from
libraries that we do not control - all the various old CDH, HDP stage
libraries, ...

The interface in Github seems much more useful to see what is going on,
so let's perhaps disable this flaky plugin completely.

Change-Id: I1547a2339e59b6154df3981af6d25320dd21a17c