SDC-11244: Add new input lane count and labels to StageDefinition structures
This patch alters the SDC code for stage extraction to work with the new
StageDefinition inputs such as number of input lanes and input lanes
labels. It's modeled after already existing output streams.

There are no validations yet, that will be a separate patch to keep
things easier to review.

Change-Id: Ie3f6f34c25da7cbf4f33129d52d34aa2c0bab622