SDC-11249. Allow to define auto-generated fields in Solr destination
This patch adds a new UI field, "Auto-generated Fields", which allows
the user to list the document fields that can be generated by
Solr. These fields are allowed to be omitted from the SDC records
without causing a validation error.

Additionally, it fixes the error messages related to record
validation, for which the list of conflicting fields was not correctly
printed.

Change-Id: I1ef59e41ac9bb393013d044640512f5ea2f05bad