SDC-11320: JDBC Origins: Improve error message when parsing of input data fails
Today the code that is handling converting ResultSet to Field have two
different catch fragments - one for IOError and second for SQLException.
I presume that the initial thinking was that the SQLException could
happen only for CLOB columns and thus we have made it issue a different
error record. Sadly as we have seen, the SQLException can really be
thrown for other reasons too and thus that exception is highly
deceiving.

Change-Id: I22a1d9c2d9680ea11f9e5e17e08a1d23d110dc68