SDC-11324: JDBC MultiTable origin can create duplicate offsets
It seems that part of our code is updating variable
startingPartitionOffsets and maxpartitionOffsets. Those variables are
however also used in the partition key, so what fundamentally happened is
that we committed offset when those parts were null, then updated them,
and committed new offset, creating second entry in the offset map.

I've tried to figure out why we are updating those and did not found a
good reason, so I simply dropped that code completely. I'm open to
feedback if I overlooked something.

Change-Id: I4f48a528375735177b86feb8ac103a3794c9465d