SDC-11330. SFTP Origin and Destination not always closing files in SFTP Server
In SSHJ, you open and close RemoteFile objects; the Input and Output Streams' close method does nothing
(the idea being that you can perform multiple operations on a file, without having to keep re-opening
it). SDC only does one type of operation on a file, so we end up not always closing the RemoteFile,
which means that the SFTP Server can hold onto a file until SDC disconnects (e.g. stop the pipeline).
We should make sure to close the RemoteFile when SDC closes the Stream.

Change-Id: Ie71dc562afeb87b5baada6beac9707be7e2a6830