SDC-11374: Reported errors in Stage execution doesn't always show up with stacktrace
When doing preview for reported errors (e.g. not stage exceptions nor
error records), we display further details when user clicks on "View
StackTrace. This link is however fully missing for exactly the same
error when actually running the pipeline.

The backend properly handles this scenario, the trouble is that the UI
piece that handles preview and pipeline run differs and it so happened
that we have missed this part in the pipeline run.

Change-Id: I07d3021b7286eda86212ce2fdc452668234d30a8