SDC-11441: JDBC Protolib: Propagate DatabaseVendor to additional places
Slightly larger patch that I hoped for, but sadly our code isn't really
good for extending in this area. The premise is to pass down the vendor
around (that we introduced recently) so that we can do few things
differently for each vendor. This patch starts by doing a custom
JDBCType.valueOf() which is generally unadvise to use since it doesn't
work with proprietary types at all (which they are plenty out there).

Change-Id: I98d15c2f6d650cc0a954b5b76708347b0eaacf4f