SDC-11454: Antenna Doctor: Migrate version check to preconditions
Realizing that we will want to plug-in this code into additional
products and offerings, it seems that tying the schema to "sdc version"
is a bit weird. This patch does two things:

* On high-level it changed semantic of the entity field. Now STAGE means
that "hey we're running on SDC" in addition to the rule be specific to
STAGE. I imagine when we add PIPELINE it will be similar (always runs in
SDC) and then APP for Control Hub.

* It moves the version checks to EL so that they are 'dynamic' - it
means whatever container version the rule is loaded in (in SDC it's SDC,
in Control Hub it's CH version).

Change-Id: Ied71d6ccf59f1e202bbb0cceb35d5c335a58bbe8