SDC-11478: Antenna Doctor: Either load main database or override file
The main use case of the override is to be able to develop a new rule in
running SDC. Having all the existing rules around and matching is
actually not really helping in that use case. Thus this patch switches
the logic - it will either load the override file or the normal file,
but never both.

Change-Id: I5178aaf43bf68af9c8a6dba1f04ce324ca3ab8f4