SDC-11491: Antenna Doctor: Replace ELs with JEXL
While working on just the second rule, I realized that the ELs are very
limited in terms of what they can do - especially on properly formatting
the summary and description. I looked around for a suitable alternative
and went though few of them - EL 3.0, Freemarker, Velocity, HTTL to name
a few. At the end, I've arrived at jexl as the best alternative. The
choosing criteria were:

* The project is alive (which commons-el is not).
* The project size and dependencies are reasonably large (and not 1.5MB
  worth of freemarker).
* The project can do both expression as well as templating (pretty much
  only jexl).

Change-Id: I72a27ed27182416476646e20d132a12ff848b419