SDC-11486: pipeline.test_jdbc_multitable_multithreaded.test_jdbc_multitable_consumer_to_jdbc is failing across the board
It indeed turns out to be my bug, but from a completely different side.
Not relevant to the change I did in JDBC Origin, but to my overly eager
optimization in Pipeline Finisher that always ended pipeline on first
batch - cause it ended pipeline even on empty batch which it should not.

Change-Id: Ie454c8c9ee89504d31141f8e164e18ef65abaea2