SDC-11523: Antenna Doctor: Introduce concept of product
This is mainly repository change - rather then having 1 repository, this
introduces concept of 'product' and 'schemaVersion' which are in fact
directories on the repository from which the main manifest.json file is
loaded. This enables us to introduce separate knowledge bases for
different products and if really needed, also different schema versions.

Change-Id: I910a78c15692b74801d7acbdabd1251ef5347138